txt_to_embeddings_job/requirements.txt
boto3
requests_aws4auth
tiktoken